requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
quart>=0.19.0
quart-discord>=2.1.0
aiohttp>=3.9.0
hypercorn>=0.16.0
